Added more data structure classes to hold database information.
I'm pushing what I last worked on a few months ago. I have decided that
it is more useful to actually let the database handle the data instead
of making objects in the code that hold the data and then push it to the
database. I am hoping this cuts down on the amount of code that I need
to write. Thus, ridding of the tedium of making over 10 builder classes
that do the same job with very minor tweaks.